+++ src/libraries/Microsoft.PowerFx.Core/Public/IPowerFxScope.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-using System;
+using System.Collections.Generic;
+using Microsoft.PowerFx.Core.Public;
+using Microsoft.PowerFx.Types;
 
 namespace Microsoft.PowerFx.Intellisense
 {
@@ -31,15 +33,10 @@ namespace Microsoft.PowerFx.Intellisense
         /// <param name="expression">The expression to convert for display.</param>
         /// <returns>The expression with display formatting applied.</returns>
         string ConvertToDisplay(string expression);
-
-        DefinitionsCheckResult CheckUserDefinedFunctions(string expression);
-
-        IIntellisenseResult Suggest(string expression, int cursorPosition, LSPExpressionMode mode);
     }
 
-    public enum LSPExpressionMode
+    internal interface IPowerFxScopeV2 : IPowerFxScope
     {
-        Default,
-        UserDefinedFunction,
+        IEnumerable<ExpressionError> GetErrors(string expression);
     }
 }
