+++ src/libraries/Microsoft.PowerFx.LanguageServerProtocol/Handlers/Notifications/OnDidOpenLanguageServerNotificationHandler.cs
@@ -1,11 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-using System;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Web;
 using Microsoft.PowerFx.Intellisense;
 using Microsoft.PowerFx.LanguageServerProtocol.Protocol;
 
@@ -38,19 +36,10 @@ namespace Microsoft.PowerFx.LanguageServerProtocol.Handlers
             didOpenParams.TextDocument.Uri,
             (scope) => 
             {
-                var dURI = new Uri(didOpenParams.TextDocument.Uri);
-                var query = HttpUtility.ParseQueryString(dURI.Query);
-                
-                if (!(query.Get("lspMode")?.ToString() is string lspModeStr &&
-                Enum.TryParse<LSPExpressionMode>(lspModeStr, true, out var lspMode)))
+                if (scope is UDFEditorContextScope scopeV2)
                 {
-                    lspMode = LSPExpressionMode.Default;
-                }
-
-                if (lspMode == LSPExpressionMode.UserDefinedFunction)
-                {
-                    var udfCheck = scope?.CheckUserDefinedFunctions(didOpenParams.TextDocument.Text);
-                    operationContext.OutputBuilder.WriteDiagnosticsNotification(didOpenParams.TextDocument.Uri, didOpenParams.TextDocument.Text, udfCheck.Errors.ToArray());
+                    var udfErrors = scopeV2?.GetErrors(didOpenParams.TextDocument.Text);
+                    operationContext.OutputBuilder.WriteDiagnosticsNotification(didOpenParams.TextDocument.Uri, didOpenParams.TextDocument.Text, udfErrors.ToArray());
                     return;
                 }
