+++ src/libraries/Microsoft.PowerFx.LanguageServerProtocol/Public/EditorContextScope.cs
@@ -11,6 +11,7 @@ using Microsoft.PowerFx.Core.Annotations;
 using Microsoft.PowerFx.Core.Functions;
 using Microsoft.PowerFx.Core.Glue;
 using Microsoft.PowerFx.Core.Parser;
+using Microsoft.PowerFx.Core.Public;
 using Microsoft.PowerFx.Core.Types;
 using Microsoft.PowerFx.Core.Utils;
 using Microsoft.PowerFx.Intellisense;
@@ -38,7 +39,16 @@ namespace Microsoft.PowerFx
             ReadOnlySymbolTable symbols = null)
         {
             return new EditorContextScope(engine, parserOptions, symbols);
-        }                  
+        }
+        
+        public static EditorContextScope CreateUDFEditorScope(
+            this Engine engine,
+            CultureInfo cultureInfo = null,
+            ReadOnlySymbolTable symbols = null,
+            bool allowSideEffects = false)
+        {
+            return new UDFEditorContextScope(engine, cultureInfo, symbols, allowSideEffects);
+        }
     }
 
     /// <summary>
@@ -46,7 +56,7 @@ namespace Microsoft.PowerFx
     ///  A scope is the context for a specific formula bar. 
     ///  This includes helpers to aide in customizing the editing experience. 
     /// </summary>
-    public sealed class EditorContextScope : IPowerFxScope, IPowerFxScopeFx2NL
+    public class EditorContextScope : IPowerFxScopeV2, IPowerFxScopeFx2NL
     {
         private readonly GuardSingleThreaded _guard = new GuardSingleThreaded();
 
@@ -67,67 +77,23 @@ namespace Microsoft.PowerFx
         // This captures the critical invariant: EditorContextScope corresponds to a formula bar where the user just types text, all other context is provided; 
         private readonly Func<string, CheckResult> _getCheckResult;
 
-        // Optional callback to check user defined functions.
-        private readonly Func<string, DefinitionsCheckResult> _checkUserDefinedFunctions;
-
         // Host can set optional hints about where this expression is used. 
         // This can feed into Fx2NL and other help services. 
         public UsageHints UsageHints { get; init; }
-
-        /// <summary>
-        /// Temporary storage for UDFs created during intellisense session. so we can add them to the symbol table for subsequent checks.
-        /// </summary>
-        private SymbolTable _localUDFs = new SymbolTable();
 
         internal EditorContextScope(
             Engine engine,
             ParserOptions parserOptions,
             ReadOnlySymbolTable symbols)
-            : this(
-                  (string expr) => new CheckResult(engine)
+            : this((string expr) => new CheckResult(engine)
                     .SetText(expr, parserOptions)
-                    .SetBindingInfo(symbols), 
-                  (string expr) => new DefinitionsCheckResult().SetText(expr, parserOptions).SetBindingInfo(ReadOnlySymbolTable.Compose(engine.UDFDefaultBindingSymbols, symbols)))
+                    .SetBindingInfo(symbols))
         {
-            _getCheckResult = (string expr) =>
-            {
-                var composed = ComposeWithLocal(symbols);
-                return new CheckResult(engine)
-                    .SetText(expr, parserOptions)
-                    .SetBindingInfo(composed);
-            };
-
-            // Factory for UDF checks: engine's default UDF binding symbols + caller symbols + local UDFs.
-            _checkUserDefinedFunctions = (string expr) =>
-            {
-                var udfBase = ReadOnlySymbolTable.Compose(engine.UDFDefaultBindingSymbols, symbols);
-                var composed = ComposeWithLocal(udfBase);
-                return new DefinitionsCheckResult()
-                    .SetText(expr, parserOptions)
-                    .SetBindingInfo(composed);
-            };
-        }
-
-        private ReadOnlySymbolTable ComposeWithLocal(ReadOnlySymbolTable baseSymbols)
-        {
-            // Avoid extra layering if nothing local was added.
-            if (_localUDFs == null || !_localUDFs.Functions.Any())
-            {
-                return baseSymbols;
-            }
-
-            return ReadOnlySymbolTable.Compose(baseSymbols, _localUDFs);
         }
 
         public EditorContextScope(Func<string, CheckResult> getCheckResult)
         {
             _getCheckResult = getCheckResult ?? throw new ArgumentNullException(nameof(getCheckResult));
-        }
-
-        public EditorContextScope(Func<string, CheckResult> getCheckResult, Func<string, DefinitionsCheckResult> checkUserDefinedFunctions)
-            : this(getCheckResult)
-        {
-            _checkUserDefinedFunctions = checkUserDefinedFunctions ?? throw new ArgumentNullException(nameof(checkUserDefinedFunctions));
         }
 
         #region IPowerFxScope
@@ -146,7 +112,7 @@ namespace Microsoft.PowerFx
             return check;
         }
 
-        string IPowerFxScope.ConvertToDisplay(string expression)
+        public string ConvertToDisplay(string expression)
         {
             var check = _getCheckResult(expression);
             var symbols = check._symbols;
@@ -160,7 +126,7 @@ namespace Microsoft.PowerFx
             return engine.GetDisplayExpression(expression, symbols, check.ParserCultureInfo);
         }
 
-        IIntellisenseResult IPowerFxScope.Suggest(string expression, int cursorPosition)
+        public virtual IIntellisenseResult Suggest(string expression, int cursorPosition)
         {
             // Suggestions just need the binding, not other things like Dependency Info or errors. 
             var check = _getCheckResult(expression);
@@ -175,7 +141,7 @@ namespace Microsoft.PowerFx
         /// Optional, can be null.
         /// </summary>
         public IServiceProvider Services { get; set; }
-         
+
         public void AddQuickFixHandlers(params CodeFixHandler[] codeFixHandlers)
         {
             this.AddQuickFixHandlers((IEnumerable<CodeFixHandler>)codeFixHandlers);
@@ -239,8 +205,8 @@ namespace Microsoft.PowerFx
                 catch (Exception e)
                 {
                     var e2 = new Exception($"Handler {handler.Key} threw {e.GetDetailedExceptionMessage()}", e);
-                    
-                        // Dont' let exceptions from a handler block other handlers. 
+
+                    // Dont' let exceptions from a handler block other handlers. 
                     logUnhandledExceptionHandler?.Invoke(e2);
                     continue;
                 }
@@ -272,198 +238,21 @@ namespace Microsoft.PowerFx
                 }
 
                 handler.OnCodeActionApplied(codeAction.ActionResultContext.ActionIdentifier);
-            }            
+            }
         }
 
         public Fx2NLParameters GetFx2NLParameters()
         {
             return new Fx2NLParameters
             {
-                 UsageHints = this.UsageHints
+                UsageHints = this.UsageHints
             };
         }
 
-        public DefinitionsCheckResult CheckUserDefinedFunctions(string expression)
+        public IEnumerable<ExpressionError> GetErrors(string expression)
         {
-            return _checkUserDefinedFunctions?.Invoke(expression) ?? throw new NotImplementedException();
-        }
-
-        public IIntellisenseResult Suggest(string expression, int cursorPosition, LSPExpressionMode mode)
-        {
-            return mode switch
-            {
-                LSPExpressionMode.Default => ((IPowerFxScope)this).Suggest(expression, cursorPosition),
-                LSPExpressionMode.UserDefinedFunction => SuggestForUdf(expression, cursorPosition),
-                _ => throw new ArgumentException($"Unknown LSP mode {mode}")
-            };
-        }
-
-        private IIntellisenseResult SuggestForUdf(string expression, int cursorPosition)
-        {
-            var checkUdf = CheckUserDefinedFunctions(expression);
-            var parsedUdfs = checkUdf.ApplyParse();
-
-            // Add valid functions to engine if not already done.
-            var funcs = checkUdf.ApplyCreateUserDefinedFunctions();
-            _localUDFs = new SymbolTable();
-            _localUDFs.AddFunctions(funcs);
-
-            // $$$ find a better way to get engine.
-            var cr = _getCheckResult.Invoke(string.Empty);
-            var engine = cr.Engine;
-
-            foreach (var parsedUdf in parsedUdfs.UDFs)
-            {
-                // Maybe cache per-UDF if this is on a hot path.
-                var udf = UserDefinedFunction.CreatePartialFunction(parsedUdf, checkUdf.UDFBindingSymbols, out _);
-                var nameResolver = udf.GetUDFNameResolver(checkUdf.UDFBindingSymbols);
-
-                // No body: only type suggestions (arg types OR return type)
-                if (parsedUdf.Body == null)
-                {
-                    // Parameter type position: foo(x: |
-                    foreach (var arg in parsedUdf.Args)
-                    {
-                        if (!IsUDFArgTypePosition(cursorPosition, arg))
-                        {
-                            continue;
-                        }
-
-                        var suggestions = BuildTypeSuggestionList(
-                            nameResolver.NamedTypes,
-                            t => !UserDefinitions.RestrictedParameterTypes.Contains(t._type));
-
-                        // start right at the type ident if present, otherwise just after the colon
-                        var startIndex = arg.TypeIdent?.Span.Min ?? (arg.ColonToken.Span.Lim + 1);
-                        var length = SafeLengthFrom(startIndex, cursorPosition);
-
-                        return MakeTypeSuggestionResult(
-                            expression,
-                            checkUdf.UDFParserOptions.Culture,
-                            suggestions,
-                            startIndex,
-                            length);
-                    }
-
-                    // Return type position: Foo(...): |
-                    if (IsUDFReturnTypePosition(cursorPosition, parsedUdf))
-                    {
-                        var suggestions = BuildTypeSuggestionList(
-                            nameResolver.NamedTypes,
-                            t => !UserDefinitions.RestrictedTypes.Contains(t._type));
-
-                        // If a return type ident already exists, replace from its start; else from after the ':' of return type.
-                        var rtStart = parsedUdf.ReturnType?.Span.Min
-                                      ?? (parsedUdf.ReturnTypeColonToken?.Span.Lim + 1)
-                                      ?? cursorPosition; // fallback safety
-
-                        var length = SafeLengthFrom(rtStart, cursorPosition);
-
-                        return MakeTypeSuggestionResult(
-                            expression,
-                            checkUdf.UDFParserOptions.Culture,
-                            suggestions,
-                            rtStart,
-                            length);
-                    }
-
-                    // Not in a typeable spot within a body-less UDF; check the next UDF.
-                    continue;
-                }
-
-                // Has body: delegate to engine if cursor is inside body span
-                var bodySpan = parsedUdf.Body.GetCompleteSpan();
-                if (cursorPosition > bodySpan.Min && cursorPosition <= bodySpan.Lim)
-                {
-                    var formula = new Formula(expression, parsedUdf.Body, checkUdf.UDFParserOptions.Culture);
-                    var binding = udf.BindBody(checkUdf.UDFBindingSymbols, new Glue2DocumentBinderGlue(), checkUdf.UDFBindingConfig);
-                    return engine.Suggest(formula, binding, cursorPosition, this.Services);
-                }
-            }
-
-            return MakeTypeSuggestionResult(
-                script: expression,
-                culture: checkUdf.UDFParserOptions.Culture, // or use default parser culture
-                suggestions: new IntellisenseSuggestionList(),
-                replacementStartIndex: cursorPosition,
-                replacementLength: 0);
-        }
-
-        private static int SafeLengthFrom(int startIndex, int cursorPosition)
-        {
-            var len = cursorPosition - startIndex;
-            return len < 0 ? 0 : len;
-        }
-
-        private static IntellisenseSuggestionList BuildTypeSuggestionList(
-            IEnumerable<KeyValuePair<DName, FormulaType>> namedTypes,
-            Func<FormulaType, bool> includePredicate)
-        {
-            var list = new IntellisenseSuggestionList();
-
-            foreach (var kvp in namedTypes)
-            {
-                var typeName = kvp.Key.Value;
-                var val = kvp.Value;
-
-                if (!includePredicate(val))
-                {
-                    continue;
-                }
-
-                list.Add(new IntellisenseSuggestion(
-                    new UIString(typeName),         // UI text
-                    SuggestionKind.Type,             // Kind
-                    SuggestionIconKind.Other,        // Icon
-                    val._type,                       // DType
-                    typeName,                       // exact match
-                    -1,                              // argCount (N/A for types)
-                    string.Empty,                    // description
-                    string.Empty));                  // help
-            }
-
-            return list;
-        }
-
-        private static IIntellisenseResult MakeTypeSuggestionResult(
-            string script,
-            CultureInfo culture,
-            IntellisenseSuggestionList suggestions,
-            int replacementStartIndex,
-            int replacementLength)
-        {
-            var data = new UDFIntellisenseData(
-                replacementStartIndex: replacementStartIndex,
-                replacementLength: replacementLength,
-                0,
-                0,
-                script: script,
-                locale: culture);
-
-            suggestions.Sort(culture);
-            IList<IntellisenseSuggestion> list = suggestions;
-
-            return new IntellisenseResult(
-                data,
-                (IReadOnlyList<IntellisenseSuggestion>)list,
-                Enumerable.Empty<TexlFunction>());
-        }
-
-        internal static bool IsUDFArgTypePosition(int cursorPos, UDFArg arg)
-        {
-            Contracts.AssertValue(arg);
-            return (arg.TypeIdent != null && arg.TypeIdent.Span.IsInRange(cursorPos)) ||
-                   (arg.ColonToken != null && arg.ColonToken.Span.IsInRange(cursorPos)) ||
-                   (arg.TypeIdent == null && arg.ColonToken != null && cursorPos > arg.ColonToken.Span.Lim);
-        }
-
-        internal static bool IsUDFReturnTypePosition(int cursorPos, UDF parsedUDF)
-        {
-            Contracts.AssertValue(parsedUDF);
-
-            return (parsedUDF.ReturnType != null && parsedUDF.ReturnType.Span.IsInRange(cursorPos)) ||
-                   (parsedUDF.ReturnTypeColonToken != null && parsedUDF.ReturnTypeColonToken.Span.IsInRange(cursorPos)) ||
-                   (parsedUDF.ReturnType == null && parsedUDF.ReturnTypeColonToken != null && cursorPos > parsedUDF.ReturnTypeColonToken.Span.Lim);
+            var check = _getCheckResult(expression);
+            return check.Errors;
         }
     }
 }
