+++ src/tests/Microsoft.PowerFx.Interpreter.Tests.Shared/IntellisenseTests/UDFIntellisenseTests.cs
@@ -19,15 +19,18 @@ namespace Microsoft.PowerFx.Core.Tests.Shared.IntellisenseTests
         [InlineData("AddNumbers(x: Number, y: Number): Number = x + Su|", "ErrorKind.InsufficientMemory,ErrorKind.NotSupported,StartOfWeek.Sunday,Sum,Substitute,TraceOptions.IgnoreUnsupportedTypes")]
         [InlineData("AddNumbers(x: Number, y: Number): |", "Boolean,Color,Date,DateTime,Dynamic,GUID,Hyperlink,Number,Text,Time,Void")]
         [InlineData("AddNumbers(x: Number, y: |", "Boolean,Color,Date,DateTime,Dynamic,GUID,Hyperlink,Number,Text,Time")]
+
+        // Suggest UDF names when calling one UDF from another
+        [InlineData("AddNumbers(x: Number, y: Number): Number = x + y; AddNumbers2(x: Number, y: Text): Number = AddNum|", "AddNumbers")]
         public void UDFSuggestionTest(string expression, string expected)
         {
             var config = new PowerFxConfig();
             var engine = new Engine(config);
-            var scope = engine.CreateEditorScope();
+            var scope = engine.CreateUDFEditorScope();
 
             // engine.AddUserDefinedFunction(expression);
 
-            var iResult = scope.Suggest(expression, expression.IndexOf('|'), LSPExpressionMode.UserDefinedFunction);
+            var iResult = scope.Suggest(expression, expression.IndexOf('|'));
 
             var suggestions = iResult.Suggestions;
